Emotion Based Smart Attendance System/Version 2/requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
requests==2.25.1
google-cloud-storage==1.42.0
werkzeug==2.0.1